pyproject.toml
[tool.setuptools]
py-modules = []

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mittaridatapumppu-deviceregistry"
description = ""
readme = "README.md"
requires-python = ">=3.10"
dynamic = ["version"]

dependencies = [
  "Django >= 4.2.6, < 5.0",
  "django-auditlog",
  "django-environ",
  "djangorestframework",
  "openpyxl",
  "Pillow",
  "psycopg2-binary",
  "pytest",
  "pytest-django",
  "sentry-sdk[django]",
  "sqlparse",
  "unittest-parametrize",
  "uvicorn",
  "whitenoise < 7",
]

[tool.black]
line-length = 120
target-version = ['py310', 'py311']
exclude = '''
(
  migrations         # migrations are automatically generated and should we excluded
  |
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.tox
    | \.venv
    | venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''
# force-exclude is for pre-commit
force-exclude = '''
(
   migrations                     # files in this dir are ignored
)
'''

[tool.ruff]
line-length = 120
target-version = "py311"
ignore-init-module-imports = true
exclude = ["migrations"]

# Enable Pyflakes `E` and `F` codes by default.
select = ["E", "F"]
ignore = []

[tool.autoflake]
in-place = true
remove-all-unused-imports =true
remove-unused-variables = true
exclude = ["migrations", "venv"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "deviceregistry.settings"
log_cli = 1
log_cli_level = 20
testpaths = ["devices/tests"]
addopts = ["--color=yes"]
